Revised members dealing specifically with local time by removing delegation to `DateTime` methods to avoid an extra call to `TimeZoneInfo` to determine offset:
> Now
> FromFileTime(long)
> ToLocalTime()