@stephentoub 2019-06-23
Can you please share performance tests and results demonstrating that this is an improvement? Thanks.