@ts2do 2019-06-23
I'm not exactly sure the best way of going about this. Can you point me to a good example?
Until then, here are some preliminary results:
`Now`: 485 ns/op >> 315 ns/op
`FromFileTime(long)`: 515 ns/op >> 300 ns/op
`ToLocalTime()`: 470 ns >> 250 ns/op