@ts2do 2019-06-30
I figured out what to do from your recent PR.

Here are the results:
**Before**

|        Method |      Mean |     Error |    StdDev |  Gen 0 |  Gen 1 |  Gen 2 |  Allocated |
| ------------- | --------- | --------- | --------- | ------ | ------ | ------ | ---------- |
|           Now |  474.7 ns |  1.913 ns |  1.790 ns |      - |      - |      - |          - |
|  FromFileTime |  498.1 ns |  1.685 ns |  1.494 ns |      - |      - |      - |          - |
|  ToLocalTime  |  459.5 ns |  1.732 ns |  1.620 ns |      - |      - |      - |          - |

**After**

|        Method |      Mean |     Error |    StdDev |  Gen 0 |  Gen 1 |  Gen 2 |  Allocated |
| ------------- | --------- | --------- | --------- | ------ | ------ | ------ | ---------- |
|           Now |  303.8 ns |  1.438 ns |  1.269 ns |      - |      - |      - |          - |
|  FromFileTime |  290.1 ns |  1.611 ns |  1.507 ns |      - |      - |      - |          - |
|   ToLocalTime |  236.7 ns |  0.881 ns |  0.824 ns |      - |      - |      - |          - |

And here is the benchmark:
```csharp
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using System;

[MemoryDiagnoser]
public class Program
{
    static void Main(string[] args) => BenchmarkSwitcher.FromTypes(new[] { typeof(Program) }).Run(args);

    private long _fileTime;
    private DateTimeOffset _utc;

    [GlobalSetup]
    public void GlobalSetup()
    {
        _utc = DateTimeOffset.UtcNow;
        _fileTime = _utc.ToFileTime();
    }

    [Benchmark] public DateTimeOffset Now() => DateTimeOffset.Now;
    [Benchmark] public DateTimeOffset FromFileTime() => DateTimeOffset.FromFileTime(_fileTime);
    [Benchmark] public DateTimeOffset ToLocalTime() => _utc.ToLocalTime();
}
```