@stephentoub 2019-07-11
@ts2do, can you please submit this as a PR in coreclr instead?  The code you're modifying here isn't actually compiled in corefx.  I assume you got the benchmark numbers then by actually making these changes in coreclr instead?