@stephentoub 2019-07-11
Per my previous comment, I'm going to close this.  A PR in coreclr would be welcome.